TINY-9754: refactor
